Fix unicode to latin-1 translation.
The UC (unicode to Latin-1) option for the tr// operator was removed in
perl 5.7. It is not possible to include the check conditionally if $] is
e.g. less than 5.008 either since the code will not compile on newer
versions of perl. I suggest we drop support for perl version 5.7 and
below by removing the check.